+++ packages/ui/button/index.ts
@@ -1,1 +1,1 @@
-export * from './button';
+export * from './cta-button';
